Split ops.array out of ops.builtin
