+++ funsor/ops/__init__.py
@@ -1,3 +1,6 @@
+# Copyright Contributors to the Pyro project.
+# SPDX-License-Identifier: Apache-2.0
+
 from .array import *
 from .builtin import *
 from .op import *
